remove section on errors due to duplicate round IDs
This section had brought up concerns [1] about the kind of scripting
behaviour this was encouraging. I have decided to remove it to move this
forward [2]. To incorporate a section in here, the narrative needs to be
changed so that it has a clear goal that a hub administrator can
identify with.

[1]: https://github.com/hubverse-org/hubAdmin/pull/86#pullrequestreview-2511387156
[2]: https://github.com/hubverse-org/hubAdmin/pull/86#issuecomment-2561241857